css/css-transforms/parsing/perspective-invalid.html

<!DOCTYPE html>
<html>
<head>
<meta charset="utf-8">
<title>CSS Transform Module Level 2: parsing perspective with invalid values</title>
<link rel="help" href="https://drafts.csswg.org/css-transforms-2/#perspective-property">
<meta name="assert" content="perspective supports only the 'none | <length>' grammar.">
<meta name="assert" content="Negative values are invalid.">
<script src="/resources/testharness.js"></script>
<script src="/resources/testharnessreport.js"></script>
<script src="/css/support/parsing-testcommon.js"></script>
</head>
<body>
<script>
test_invalid_value("perspective", "auto");
test_invalid_value("perspective", "20%");
test_invalid_value("perspective", "-30px");
test_invalid_value("perspective", "40");
test_invalid_value("perspective", "calc(50% + 60px)");
test_invalid_value("perspective", "70px 80px");
test_invalid_value("perspective", "none 90px");
test_invalid_value("perspective", "calc(0)");
</script>
</body>
</html>
